Allow compilation to succeed on OSX
Fix a few instances of unportable Code:

 - make the get_current_dir_name() call dependant on a configure check

 - make time logging code use casts to long uniformly

 - include some headers where needed

 - replace assignment of '\0' with NULL so modern clang doesn't complain
   about assigning null pointers via expressions evaluating to zero

+++ naemon/logging.c
@@ -438,7 +438,7 @@ int log_debug_info(int level, int verbosity, const char *fmt, ...)
 
 	/* write the timestamp */
 	gettimeofday(&current_time, NULL);
-	fprintf(debug_file_fp, "[%lu.%06lu] [%03d.%d] [pid=%lu] ", current_time.tv_sec, current_time.tv_usec, level, verbosity, (unsigned long)getpid());
+	fprintf(debug_file_fp, "[%ld.%06ld] [%03d.%d] [pid=%lu] ", (long)current_time.tv_sec, (long)current_time.tv_usec, level, verbosity, (unsigned long)getpid());
 
 	/* write the data */
 	va_start(ap, fmt);
